fix: layout on project board, move directory bar out of overscroll zone
