+++ client/components/ProjectView.js
@@ -2,6 +2,7 @@ import React from 'react'
 import {withRouter} from 'react-router'
 import ProjectProvider from '../context/projectContext'
 import Board from './Board'
+import DirectoryBar from './DirectoryBar'
 import styles from './css/ProjectView.css'
 
 const ProjectView = ({match}) => {
@@ -10,6 +11,7 @@ const ProjectView = ({match}) => {
   return (
     <div className={styles.projectContainer}>
       <ProjectProvider projectId={projectId}>
+        <DirectoryBar />
         <Board />
       </ProjectProvider>
     </div>
