+++ client/components/css/AddColumnDropDown.css
@@ -3,11 +3,12 @@
   flex-direction: column;
   align-items: center;
   justify-content: flex-start;
-  width: 320px;
+  width: 355px;
   margin: 0 0.3em;
-  height: 100vh;
+  height: 510px;
   background-color: rgb(248, 248, 248);
   border: 4px dotted darkgrey;
+  border-radius: 5px;
   padding: 0.5em;
 }
