+++ client/components/css/ProjectView.css
@@ -1,7 +1,7 @@
 .projectContainer {
-  height: 800px;
-  width: 100%;
-  margin: 0 auto;
-  padding: 0;
-  overflow-x: scroll;
+  display: flex;
+  flex-direction: column;
+  width: 100vw;
+  height: 530px;
+  padding: 0.5em;
 }
